pyproject.toml
[tool.poetry]
name = "phoenix-bot"
version = "0.1.0"
description = "aiogram bot for phoenix studio"
authors = ["macalistervadim <macalistervadim@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.6.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-absolute-import = "^1.0.0.2"
flake8-commas = "^2.1.0"
flake8-import-order = "^0.18.2"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
flake8-use-pathlib = "^0.3.0"
pep8-naming = "^0.13.3"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
